reordenamiento de redux. fino fino
