+++ Client/src/Redux/Features/categories/clientCategoriesSlice.js
@@ -31,7 +31,8 @@ const clientCategoriesSlice = createSlice({
             })
             .addCase(getAllCategories.fulfilled, (state, action) => {
                 state.status = 'succeeded',
-                state.allCategories = action.payload
+                state.allCategories = action.payload,
+                state.error = null
             })
             .addCase(getAllCategories.rejected, (state, action) => {
                 state.status = 'rejected',
