+++ Client/src/components/Client/ContainerAllProducts/ContainerAllProducts.jsx
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux';
 import { Pagination, Row, Col, Empty, Space } from 'antd';
-import { getAllProducts } from '../../../Redux/Features/productsClient/productsClientSlice';
+import { getAllProducts } from '../../../Redux/Features/products/clientProductsSlice';
 import ProductCard from '../ProductCard/ProductCard';
 
 const ContainerAllProducts = () => {
 
     const dispatch = useDispatch()
-    const allProducts = useSelector((state) => state.productsClient.allProducts)
+    const allProducts = useSelector((state) => state.clientProducts.allProducts)
     const categoryName = window.localStorage.getItem('category_name');
     //console.log('estoy en conteinerProduct', allProducts)
