+++ Client/src/components/Client/ConteinerProductsByCategory/ConteinerProductsByCategory.jsx
@@ -1,14 +1,14 @@
 import React, { useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux';
 import { Pagination, Row, Col, Empty, Space } from 'antd';
-import { getProductsByCategory, clearProductsByCategory} from '../../../Redux/Features/productsClient/productsClientSlice';
+import { getProductsByCategory, clearProductsByCategory} from '../../../Redux/Features/products/clientProductsSlice';
 import ProductCard from '../ProductCard/ProductCard';
 // import { useParams } from 'react-router-dom';
 
 const ConteinerProductsByCategory = () => {
 
     const dispatch = useDispatch()
-    const productsByCategory = useSelector( state => state.productsClient.productsByCategory)
+    const productsByCategory = useSelector( state => state.clientProducts.productsByCategory)
     const idProduct = window.localStorage.getItem('category_id')
 
     //console.log('idparams en productcategori componente', idProduct)
