+++ Client/src/components/Client/DashboardUser/DashboardUser.jsx
@@ -3,9 +3,6 @@ import { Row, Col } from 'antd';
 import Contacticons from "./Contacticons/Contacticons"
 import MenuDash from "./MenuDash/MenuDash";
 import Nav from "./Nav/Nav";
-import { useEffect } from "react";
-import { useDispatch, useSelector } from "react-redux";
-import { getItems } from "../../../Redux/Features/productsClient/productsClientSlice";
 
 const DashboardUser = () => {
     return (
